Update README.md
This Project is build using HTML,CSS and Javascript for frontend and Node.js and it's frame works for backend. MongoDb is used as the Database which is located at MongoDB atlas instead of local machine.

1. Home page
2. Login page
3. Sign Up page
4. Index page: This is the page to show all the campgrounds present in the database.
5. Show page: This page shows details of a particular campground like location on map, reviews and pictures.
6. Update page: This page lets the authorized user to edit the details of that particular campground.

Authentication is also there to see if the details like username and password entered are all correct and present in the database.
Authorization is also there which only gives the user the authority to edit or delete a campground if and only if it is uploaded by the same person.